backend/requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.16,<3.17
djangorestframework-simplejwt>=5.5,<5.6
django-cors-headers>=4.9,<5.0
dj-database-url>=3.0,<4.0
gunicorn>=23.0,<24.0
Pillow>=11.0,<13.0
psycopg[binary]>=3.2,<3.3
python-dotenv>=1.0,<2.0
whitenoise>=6.9,<7.0
